Headline: Add commit_message command
This adds a new command to GPT3.py that requests some user input for generating a commit message for a give diff. The user's text will be used as a prompt for GPT-3, and all of the logic for collecting a response from the user and prompting GPT is abstracted in the base_query helper function.