Refactor GPT3 response handling to support streaming output
This commit modifies the `base_query` function in `gpt3.py` to add support for streaming output of GPT3 responses. It does this by using the `stream` parameter in the `openai.ChatCompletion.create` call and then processing the resulting `chunk` objects in a loop, appending the response content to the appropriate index in a list. If `stream_output` is True, the function returns early to allow the streaming to take place. Otherwise, the response contents are printed to stdout in the same format as before.