feat(commit): Run LLMs in parallel and display execution time
* To improve the performance of the `build_commit` command by running language
models (LLMs) in parallel.
* To provide users with more information about the execution time of each LLM.

* Added a new function `async_run_on_llms` that takes a function and a list of
LLMs as input and runs the function on each LLM in parallel.
* The `async_run_on_llms` function also measures the execution time of each LLM
and returns a list of results that includes the output of the function, the LLM
that was used, and the execution time.
* Modified the `a_build_commit` function to use the `async_run_on_llms` function
to run the `describe_diff` function on each LLM in parallel.
* Updated the output of the `build_commit` command to include the execution time
of each LLM.